Dia4/email-service/requirements.txt
Django==5.0
djangorestframework==3.15
django-cors-headers
psycopg2-binary
redis
celery==5.4
dj-database-url